more granularity of latency

+++ crates/orchestrator/src/lib.rs
@@ -479,9 +479,9 @@ where
                 // Deal with the bench results from different nodes
                 let cur_metrics = self.bench_results.clone();
                 self.bench_results.avg_latency_in_sec = (metrics.avg_latency_in_sec
-                    * metrics.num_latency
-                    + cur_metrics.avg_latency_in_sec * cur_metrics.num_latency)
-                    / (metrics.num_latency + cur_metrics.num_latency);
+                    * (metrics.num_latency as f64)
+                    + cur_metrics.avg_latency_in_sec * (cur_metrics.num_latency as f64))
+                    / (metrics.num_latency + cur_metrics.num_latency) as f64;
                 self.bench_results.num_latency += metrics.num_latency;
                 self.bench_results.minimum_latency_in_sec = metrics
                     .minimum_latency_in_sec
